requirements.txt
pygame>=2.5.0
pyyaml>=6.0
feedparser>=6.0.0
numpy>=1.20.0
PyOpenGL>=3.1.0
# For NTN24 scraper
requests>=2.28.0
beautifulsoup4>=4.11.0
# For Syphon support on macOS
#syphon-python>=0.1.0
# For NDI support (cross-platform)
# Note: NDI requires additional setup - see README for installation instructions
# ndi-python>=5.0.0  # Commented out due to build requirements
opencv-python>=4.5.0
# For Spout support on Windows (much easier than NDI)
SpoutGL>=0.1.0
